refactor: modify now-playing widget
